+++ CMakeLists.txt
@@ -2,4 +2,5 @@ cmake_minimum_required(VERSION 3.10)
 
 project(Reversi-AI VERSION 0.1)
 
+include(common/CMakeLists.txt)
 include(visualizer/CMakeLists.txt)
