+++ visualizer/Game.h
@@ -4,7 +4,7 @@
 #include <array>
 #include <utility>
 
-#include "Desk.h"
+#include "common/Desk.h"
 #include "Player.h"
 #include "WindowPlayer.h"
