+++ visualizer/WindowPlayer.h
@@ -4,7 +4,7 @@
 #include <utility>
 #include <condition_variable>
 
-#include "Desk.h"
+#include "common/Desk.h"
 #include "Player.h"
 
 class WindowPlayer :
